changing converting the types of the termination_condition and solution_status attributes on the SolverResults object to simple strings. The Enums are still used internally and for documentation
